requirements.txt
flask==2.0.1
pillow==7.0.0
pymongo==3.12.0
ipfshttpclient==0.8.0a2